Fixing various pieces of the testing framework and other general cleanup.
-Mikeal



git-svn-id: http://svn.getwindmill.com/trunk@479 78c7df6f-8922-0410-bcd3-9426b1ad491b

+++ windmill/browser/firefox.py
@@ -197,7 +197,7 @@ class MozillaBrowser(object):
                 
         if windmill.settings['MOZILLA_REMOVE_PROFILE_ON_EXIT'] is True:
             # Windows holds on to the file handlers for prefs.js indefinitely, we leave tempfiles and let the OS handle cleaning them up at a later time 
-            if sys.platform != "win32":
+            if sys.platform != "win32" and os.path.isdir(windmill.settings['MOZILLA_PROFILE_PATH']):
                 shutil.rmtree(windmill.settings['MOZILLA_PROFILE_PATH'])
 
     def stop(self):
